requirements.txt
abcplus==0.1.0
alembic==1.4.2
click==6.7
DateTime==4.3
Flask==1.1.2
Flask-Jsonpify==1.5.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.3.2
gunicorn==20.0.4
itsdangerous==0.24
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.0
mysqlclient==1.4.6
numpy==1.19.0
passlib==1.7.2
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.2.12
Werkzeug==1.0.1
zope.interface==5.1.0
